[SCV-26] Cleaned up validator directory I split the validator functions for the schema and stac into seperate files exported thorugh the index file
